GuideView 주석 추가
GuideView 상세한 주석 넣는중

+++ Noo6Main/DataStructS/DataStruct.swift
@@ -45,7 +45,7 @@ struct GuideInfo: Codable, Identifiable{
     var id : Int
     var guideView : [String]
     var guideMessage : [String]
-    var totalPage : Double
+    var totalPage : Int
     var voice : [String]
 }
